Move orders functionality from App to Orders, make App a functional component
