+++ src/containers/App/App.js
@@ -7,41 +7,17 @@ import { getOrders } from '../../apiCalls';
 import Orders from '../../components/Orders/Orders';
 import OrderForm from '../../components/OrderForm/OrderForm';
 
-export class App extends Component {
-  constructor(props) {
-    super();
-    this.props = props;
-  }
+export const App = () => {
 
-  componentDidMount() {
-    getOrders()
-      .then(data => this.props.setOrders(data.orders))
-      .catch(err => console.error('Error fetching:', err));
-  }
-
-  render() {
     return (
       <main className="App">
         <header>
           <h1>Burrito Builder</h1>
           <OrderForm />
         </header>
-        
-        <Orders orders={this.props.orders}/>
       </main>
     );
   }
-}
 
-export const mapStateToProps = ({ orders }) => ({
-  orders
-});
-
-export const mapDispatchToProps = dispatch => (
-  bindActionCreators({
-    setOrders,
-  }, dispatch)
-);
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
